+++ src/docs/views/components/fragments/component-page/fragments/prop-title/prop-title.vue
@@ -1,24 +1,30 @@
 <template>
-	<section-title v-bind="{ id }" class="flex flex-wrap gap-4 items-baseline relative group/prop-title">
+	<section-title
+		v-bind="{ id }"
+		class="group/prop-title relative flex flex-wrap items-baseline gap-4"
+	>
 		<template #title>
 			<slot name="title" />
 		</template>
 
 		<template #post-title>
-			<dl v-if="haveType || haveDefault" class="flex not-prose text-sm border border-grey-200 rounded dark:border-transparent dark:bg-white/20">
-				<div v-if="haveType" class="flex flex-wrap items-center gap-2 px-2 py-1 items-start">
-					<dt class="font-bold">
-						Type
-					</dt>
+			<dl
+				v-if="haveType || haveDefault"
+				class="not-prose border-grey-200 flex rounded border text-sm dark:border-transparent dark:bg-white/20"
+			>
+				<div v-if="haveType" class="flex items-start gap-2 px-2 py-1">
+					<dt class="font-bold">Type</dt>
 					<dd class="text-grey-500 dark:text-white/60">
 						<slot name="type" />
 					</dd>
 				</div>
-				<div v-if="haveDefault" class="flex flex-wrap items-center gap-2 px-2 py-1 items-start" :class="{ 'border-s border-grey-200 dark:border-transparent': haveType }">
-					<dt class="font-bold">
-						Default
-					</dt>
-					<dd class="text-grey-500 dark:text-white/60 font-mono text-balance">
+				<div
+					v-if="haveDefault"
+					class="flex shrink-0 items-start gap-2 px-2 py-1"
+					:class="{ 'border-grey-200 border-s dark:border-transparent': haveType }"
+				>
+					<dt class="font-bold">Default</dt>
+					<dd class="text-grey-500 font-mono text-balance dark:text-white/60">
 						<slot name="default-value" />
 					</dd>
 				</div>